feat(greeks): add GreeksMonitor service with repository
Add GreeksMonitor service (Tasks 20-23) that orchestrates Greeks
monitoring including calculation, aggregation, alerting, and persistence.

Key components:
- GreeksRepository: Persists snapshots and alerts to database
  - save_snapshot: Persist AggregatedGreeks to database
  - get_latest_snapshot: Get most recent snapshot for a scope
  - get_prev_snapshot: Get snapshot from ~N seconds ago for ROC
  - save_alert/acknowledge_alert: Manage alerts in database
  - get_unacknowledged_alerts: Query pending alerts

- GreeksMonitor: Orchestrates full monitoring cycle
  - Coordinates Calculator, Aggregator, AlertEngine, Repository
  - check(): Run full monitoring cycle with alert generation
  - get_current_greeks(): Return cached account Greeks

- Helper functions:
  - load_positions_from_db: Load option positions from DB as PositionInfo
  - create_greeks_monitor: Factory for creating configured monitors

Also fixes BigInteger primary key issue in greeks.py models for SQLite
compatibility by using Integer instead.

Note: Pre-existing test failure in tests/options/test_checker.py
unrelated to these changes.

+++ backend/src/models/greeks.py
@@ -5,7 +5,7 @@ import uuid
 from datetime import datetime
 from decimal import Decimal
 
-from sqlalchemy import BigInteger, Boolean, DateTime, Integer, Numeric, String, Text
+from sqlalchemy import Boolean, DateTime, Integer, Numeric, String, Text
 from sqlalchemy.dialects.postgresql import UUID as PGUUID
 from sqlalchemy.orm import Mapped, mapped_column
 
@@ -21,7 +21,7 @@ class GreeksSnapshot(Base):
 
     __tablename__ = "greeks_snapshots"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
 
     # Scope identification
     scope: Mapped[str] = mapped_column(String(20), nullable=False)  # 'ACCOUNT' or 'STRATEGY'
@@ -56,7 +56,7 @@ class GreeksAlertRecord(Base):
 
     __tablename__ = "greeks_alerts"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     alert_id: Mapped[uuid.UUID] = mapped_column(PGUUID(as_uuid=True), unique=True, nullable=False)
 
     # Alert classification
